Create Python If-Else
Task
Given an integer, , perform the following conditional actions:

If  is n odd, print Weird
If  is n even and in the inclusive range of  to , print Not Weird
If  is n even and in the inclusive range of  to , print Weird
If  is  n even and greater than , print Not Weird
Input Format

A single line containing a positive integer, .

Constraints

Output Format

Print Weird if the number is weird. Otherwise, print Not Weird.